refactor(hooks): extract focused room hooks; migrate modals + RoomView (Phase 2)
The room action callbacks were defined verbatim in three hooks (useRoom,
useRoomActions, useRoomActive). This extracts the slice-shaped ones into
focused, render-stable hooks and makes useRoomActions compose them.

- New hooks (each subscribes to NO store): usePolls (send/vote/close),
  useRoomModeration (moderation + XEP-0045 affiliations/roles + XEP-0317
  hats), useRoomManagement (room CRUD/config/bookmarks/invites/avatars).
- useRoomActions composes the three (−~280 lines); its full action surface
  is unchanged, so useRoom (which composes useRoomActions) is unaffected.
- Exported from all three barrels.
- Consumer migrations: RoomMembersModal + RoomHatsModal drop from the full
  useRoomActions to useRoomModeration; RoomView sources its poll /
  moderation / management slices from the focused hooks instead of the
  useRoomActive spread.

No render churn reintroduced: a render-stability test asserts each focused
hook AND the composed useRoomActions keep a STABLE object identity across
heavy room-store churn (they subscribe to nothing). useRoomActive is left
untouched (message-list-adjacent); deduping it via the focused hooks is a
follow-up.
