+++ packages/fluux-sdk/src/hooks/index.ts
@@ -10,6 +10,9 @@ export { useConsole } from './useConsole'
 export { useEvents } from './useEvents'
 export { useRoom } from './useRoom'
 export { useRoomActive } from './useRoomActive'
+export { usePolls } from './usePolls'
+export { useRoomModeration } from './useRoomModeration'
+export { useRoomManagement } from './useRoomManagement'
 export { useXMPP } from './useXMPP'
 export { useBlocking } from './useBlocking'
 export { useIgnore } from './useIgnore'
