+++ packages/fluux-sdk/src/index.ts
@@ -78,6 +78,9 @@ export { useEvents } from './hooks/useEvents'
 export { useRoom } from './hooks/useRoom'
 export { useRoomActive } from './hooks/useRoomActive'
 export { useRoomActions } from './hooks/useRoomActions'
+export { usePolls } from './hooks/usePolls'
+export { useRoomModeration } from './hooks/useRoomModeration'
+export { useRoomManagement } from './hooks/useRoomManagement'
 export { useReferencedMessage, type ReferencedMessageParams } from './hooks/useReferencedMessage'
 export { useXMPP } from './hooks/useXMPP'
 export { useAdmin } from './hooks/useAdmin'
